Sécurisation + control type Utilisateur

+++ res/php/deleteSubject.php
@@ -4,6 +4,10 @@ session_start();
 $idSujet = htmlspecialchars($_GET['idSujet']);
 $from = htmlspecialchars($_GET['from']);
 
+if (!isset($_SESSION['utilisateur']) || !($_SESSION['utilisateur'] instanceof Utilisateur)) {
+    header("Location: ./index.php");
+    exit();
+} 
 $co = mysqli_connect(DB_HOST, DB_USERNAME, DB_PASSWORD, DB_NAME);
 if ($_SESSION['utilisateur']->getRole()) {
     $req = $co->query("DELETE FROM SUJET WHERE idSujet = $idSujet");
